requirements.txt
# Discord Bot
discord.py
python-dotenv
tiktoken

# Database
aiomysql
pandas
numpy

# Image Processing
Pillow

# AI/ML
openai

# Utilities
pytz
